+++ Engine/src/Editor/Panel/Panel.cpp
@@ -1,6 +1,6 @@
 #include "Panel.h"
 
-void Panel::SetFPS(float x) { vpPanel.fps = x; }
+void Panel::SetFPS(int x) { vpPanel.fps = x; }
 
 Panel::Panel(const char* t) :
